Question on index for finite approx
Hey Alex, 
Wasn't sure how to pull a request on ICT notes to ask questions, hoping this works. 

2 questions on the finite differences: 

Why is the delta x one changing j instead of i? It was originally i during the W11 ICT but then you realised it was a mistake and changed it to j for your fixing index in your commit. 

I thought the x-axis was associated with the row which is indexed with i & the y-axis uses j? Why does it switch? 

Q1: Since it's a central difference finite differences formula, is the numerator of the finite difference 2*delta(x) or 2*delta(y)